+++ application/configuration/ftp.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Created by TeraBoxX.
+ * User: tommy
+ * Date: 2/5/2017
+ * Time: 12:33 AM
+ */
+
+
+$config["ftp"]= array(
+	/*
+	
+	"default" => array(
+
+		"type" => "ftp",
+		"host" => "127.0.0.1",
+		"port" => 21,
+		"username" => "",
+		"password" => "",
+		"basedir" => ""
+	),
+	
+	*/
+);