+++ terablaze/core/Exception.php
@@ -4,6 +4,6 @@ namespace TeraBlaze\Core
 {
     class Exception extends \Exception
     {
-        
+
     }
 }