+++ terablaze/libraries/ftp/Exception.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace TeraBlaze\Libraries\Ftp
+{
+    use TeraBlaze\Core as Core;
+    
+    class Exception extends Core\Exception
+    {
+        
+    }
+}