+++ terablaze/libraries/ftp/exception/Argument.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace TeraBlaze\Libraries\Ftp\Exception
+{
+    use TeraBlaze\Core as Core;
+
+    class Argument extends Core\Exception\Argument
+    {
+
+    }
+}