+++ terablaze/libraries/ftp/exception/Service.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace TeraBlaze\Libraries\Ftp\Exception
+{
+    use TeraBlaze\Libraries\Ftp as Ftp_driver;
+    
+    class Service extends Ftp_driver\Exception
+    {
+    
+    }
+}